Deprecate gen_domain (submesh stuff)
This only works on serial, and so is only really relevant for the
ismipc test case. And it has no effect on the ismipc case because the
whole domain is valid. Therefore it only serves to mess up the
DofMaps.

For example, when viewing eigenfunctions, if loading mesh from file
and creating function spaces, the dofmap is different from in the
run_eig stage because of the use of SubMesh.