Some initial work on ice_stream test case
Not totally happy with some of this. The ice_stream case has the mesh
already generated, whereas the ismipc cases generate them on the
fly. Does this matter?

Also, the bed elev functions are defined based on case name "ismip" vs
"ice_stream". Would be better to incorporate into the [testing]
section of the TOML file, but how would this work?